move cmd/operator-sdk/internal/genutil/* -> internal/genutil/*
