+++ internal/genutil/crds.go
