+++ internal/genutil/genutil.go
